FRONT END/requirements.txt
Flask                        3.0.3
mysql-connector-python       8.3.0
numpy                        1.26.0
pandas                       2.2.2
scikit-learn                 1.3.0